`trusted.txt` is only a Windows thing
